backend/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0